@@ -2,11 +2,13 @@
 // Licensed under the MIT License.
 import * as React from 'react';
 
-export type ReactFCWithDisplayName<P = {}> = React.FC<P> & { displayName: string };
+export type ReactFCWithDisplayName<P = {}> = React.FC<React.PropsWithChildren<P>> & {
+    displayName: string;
+};
 
 export function NamedFC<P = {}>(
     displayName: string,
-    component: React.FC<P>,
+    component: React.FC<React.PropsWithChildren<P>>,
 ): ReactFCWithDisplayName<P> {
     component.displayName = displayName;
 
