@@ -2,4 +2,4 @@
 // Licensed under the MIT License.
 import { ILinkProps } from '@fluentui/react';
 
-export type LinkComponentType = React.FC<ILinkProps>;
+export type LinkComponentType = React.FC<React.PropsWithChildren<ILinkProps>>;
