@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 import '@testing-library/jest-dom';
 import { ActionButton } from '@fluentui/react';
-import { render } from '@testing-library/react';
+import { act, render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { Assessments } from 'assessments/assessments';
 import { AssessmentsProvider } from 'assessments/types/assessments-provider';
@@ -383,7 +383,7 @@ describe('DetailsViewCommandBar', () => {
             name: 'Start over Assessment',
         });
         expect(renderResult.asFragment()).toMatchSnapshot('start assessment over dialog hidden');
-        await userEvent.click(startOverAssessmentButton);
+        await userEvent.click(startOverAssessmentButton, { pointerEventsCheck: 0 });
         expect(getMockComponentClassPropsForCall(StartOverDialog, 2).dialogState).toBe(
             'assessment',
         );
@@ -438,7 +438,13 @@ describe('DetailsViewCommandBar', () => {
             await userEvent.click(exportButton); //open the dialog
             const textArea = renderResult.getByRole('textbox');
             expect(textArea).toHaveFocus();
-            getMockComponentCall(ExportDialog, 2)[0].afterDismissed();
+
+            const exportDialogProps = getMockComponentCall(ExportDialog, 2)[0];
+
+            act(() => {
+                exportDialogProps.afterDismissed();
+            });
+
             expect(exportButton).toHaveFocus();
         });
 
@@ -491,7 +497,10 @@ describe('DetailsViewCommandBar', () => {
             const props = getProps(['StartOverComponent', 'CommandBar']);
             const renderResult = render(<DetailsViewCommandBar {...props} />);
             expect(renderResult.baseElement).toHaveFocus();
-            getMockComponentCall(StartOverDialog)[0].dismissDialog();
+            act(() => {
+                getMockComponentCall(StartOverDialog)[0].dismissDialog();
+            });
+
             expect(renderResult.baseElement).toHaveFocus();
         });
 
@@ -506,16 +515,22 @@ describe('DetailsViewCommandBar', () => {
             const startOverMenuButton = renderResult.getByRole('button', {
                 name: 'start over menu',
             });
-            getMockComponentCall(StartOverDialog)[0].dismissDialog();
+            act(() => {
+                getMockComponentCall(StartOverDialog)[0].dismissDialog();
+            });
+
             expect(renderResult.baseElement).toHaveFocus();
             await userEvent.click(startOverMenuButton);
             const startOverAssessmentButton = renderResult.getByRole('menuitem', {
                 name: 'Start over Assessment',
             });
-            await userEvent.click(startOverAssessmentButton);
+            await userEvent.click(startOverAssessmentButton, { pointerEventsCheck: 0 });
             const cancelDialogButton = renderResult.getByRole('button', { name: 'Cancel' });
             expect(cancelDialogButton).toHaveFocus();
-            getMockComponentCall(StartOverDialog, 3)[0].dismissDialog();
+
+            act(() => {
+                getMockComponentCall(StartOverDialog, 3)[0].dismissDialog();
+            });
             expect(startOverMenuButton).toHaveFocus();
         });
     });
