@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { Dialog, DialogFooter } from '@fluentui/react';
-import { render } from '@testing-library/react';
+import { act, render } from '@testing-library/react';
 import { ToolData } from 'common/types/store-data/unified-data-interface';
 import * as React from 'react';
 import { IMock, It, Mock, MockBehavior, Times } from 'typemoq';
@@ -26,6 +26,7 @@ import {
     expectMockedComponentPropsToMatchSnapshots,
     getMockComponentClassPropsForCall,
     mockReactComponents,
+    mockReactComponent,
 } from '../../../mock-helpers/mock-module-helpers';
 jest.mock('@fluentui/react');
 jest.mock('../../../../../DetailsView/components/action-and-cancel-buttons-component');
@@ -34,9 +35,9 @@ describe('IssueFilingDialog', () => {
     mockReactComponents([
         ActionAndCancelButtonsComponent,
         IssueFilingSettingsContainer,
-        Dialog,
         DialogFooter,
     ]);
+    mockReactComponent(Dialog, 'Dialog');
     let eventStub: EventStub;
     let isSettingsValidMock: IMock<Function>;
     let getSettingsFromStoreDataMock: IMock<Function>;
@@ -129,7 +130,7 @@ describe('IssueFilingDialog', () => {
         const renderResult = render(<IssueFilingDialog {...props} />);
 
         expect(renderResult.asFragment()).toMatchSnapshot();
-        expectMockedComponentPropsToMatchSnapshots([Dialog]);
+        expectMockedComponentPropsToMatchSnapshots([Dialog], 'Dialog props');
     });
 
     it('render: validate correct callbacks to ActionAndCancelButtonsComponent (file issue on click and cancel)', () => {
@@ -141,6 +142,7 @@ describe('IssueFilingDialog', () => {
         );
 
         actionCancelButtons.cancelButtonOnClick(null);
+
         isSettingsValidMock.verifyAll();
         onCloseMock.verifyAll();
     });
@@ -207,10 +209,12 @@ describe('IssueFilingDialog', () => {
             propertyName: propertyStub,
             propertyValue: propertyValueStub,
         };
-        issueFilingSettingsContainer.onPropertyUpdateCallback(payload);
+        act(() => {
+            issueFilingSettingsContainer.onPropertyUpdateCallback(payload);
+        });
 
         expect(renderResult.asFragment()).toMatchSnapshot();
-        expectMockedComponentPropsToMatchSnapshots([Dialog]);
+        expectMockedComponentPropsToMatchSnapshots([Dialog], 'Dialog props');
     });
 
     it('render: validate callback (onPropertyUpdateCallback) sent to settings container when service settings are not null', () => {
@@ -235,10 +239,12 @@ describe('IssueFilingDialog', () => {
             propertyName: propertyStub,
             propertyValue: propertyValueStub,
         };
-        issueFilingSettingsContainer.onPropertyUpdateCallback(payload);
+        act(() => {
+            issueFilingSettingsContainer.onPropertyUpdateCallback(payload);
+        });
 
         expect(renderResult.asFragment()).toMatchSnapshot();
-        expectMockedComponentPropsToMatchSnapshots([Dialog]);
+        expectMockedComponentPropsToMatchSnapshots([Dialog], 'Dialog props');
     });
 
     it('render: validate callback (onSelectedServiceChange) sent to settings container', () => {
@@ -271,10 +277,12 @@ describe('IssueFilingDialog', () => {
         const payload = {
             issueFilingServiceName: differentServiceKey,
         };
-        issueFilingSettingsContainer.onSelectedServiceChange(payload);
+        act(() => {
+            issueFilingSettingsContainer.onSelectedServiceChange(payload);
+        });
 
         expect(renderResult.asFragment()).toMatchSnapshot();
-        expectMockedComponentPropsToMatchSnapshots([Dialog]);
+        expectMockedComponentPropsToMatchSnapshots([Dialog], 'Dialog props');
     });
 
     const scenarios = [
@@ -306,7 +314,7 @@ describe('IssueFilingDialog', () => {
 
             renderResult.rerender(<IssueFilingDialog {...newProps} />);
             expect(renderResult.asFragment()).toMatchSnapshot();
-            expectMockedComponentPropsToMatchSnapshots([Dialog]);
+            expectMockedComponentPropsToMatchSnapshots([Dialog], 'Dialog props');
         },
     );
 });
