@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { render } from '@testing-library/react';
+import { act, render } from '@testing-library/react';
 import * as React from 'react';
 import {
     withStoreSubscription,
@@ -10,7 +10,9 @@ import { ClientStoresHub } from '../../../../../common/stores/client-stores-hub'
 
 describe('withStoreSubscription', () => {
     type testProps = WithStoreSubscriptionProps<{ message: string }>;
-    const testComp: React.FC<testProps> = props => <h1>{props.storeState.message}</h1>;
+    const testComp: React.FC<React.PropsWithChildren<testProps>> = props => (
+        <h1>{props.storeState.message}</h1>
+    );
 
     test('constructor: storesHub is null', () => {
         const props: testProps = {
@@ -181,7 +183,7 @@ describe('withStoreSubscription', () => {
         expect(listenerAdded).toEqual(listenerRemoved);
     });
 
-    test('render', () => {
+    test('render', async () => {
         let onStoreChange;
         const hasStoresMock = jest.fn();
         const getStoreDataMock = jest.fn();
@@ -214,7 +216,7 @@ describe('withStoreSubscription', () => {
 
         expect(renderResult.container).toMatchSnapshot('before store change');
 
-        onStoreChange();
+        await act(() => onStoreChange());
 
         expect(renderResult.container).toMatchSnapshot('after store change');
     });
