@@ -3,10 +3,11 @@
 import { render } from '@testing-library/react';
 import { productName } from 'content/strings/application';
 import * as React from 'react';
-import { Helmet } from 'react-helmet';
+import { Helmet } from 'react-helmet-async';
 import { mockReactComponents } from 'tests/unit/mock-helpers/mock-module-helpers';
 import { GuidanceTitle } from 'views/content/guidance-title';
-jest.mock('react-helmet');
+
+jest.mock('react-helmet-async');
 
 describe('guidance title', () => {
     mockReactComponents([Helmet]);
